Placed actions in their own component
