+++ src/components/books/LibraryActions.module.css
@@ -1,0 +1,47 @@
+/* Actions */
+
+.actions {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 1rem;
+    justify-content: center;
+    align-items: center;
+    border-top: 1px solid gainsboro;
+    padding-top: 0.875rem;
+    cursor: pointer;
+}
+
+.actions p {
+    padding: 8px 25px;
+    font-weight: 600;
+    color: var(--dark-blue);
+    border-radius: 50px;
+    border: 2px solid var(--dark-blue);
+    width: 100%;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.actions p:hover {
+    background-color: var(--dark-blue);
+    color: var(--white);
+    border-radius: 50px;
+    padding: 8px 25px;
+}
+
+.actions p span {
+    padding-right: 5px;
+}
+
+/* Media Queries */
+@media (min-width: 450px) {
+    .actions {
+        justify-content: space-evenly;
+    }
+
+    .actions p {
+        width: 45%;
+    }
+}
