+++ src/components/books/Summary.module.css
@@ -1,19 +1,3 @@
-.info {
-  position: fixed;
-  bottom: 10px;
-  left: 0px;
-  width: 100%;
-}
-
-.book-summary {
-  background-color: var(--white);
-  width: 95%;
-  max-width: 600px;
-  margin: 0 auto;
-  border-radius: 10px;
-  padding: 1rem;
-}
-
 .summary {
   display: grid;
   grid-template-columns: 100px auto;
@@ -80,57 +64,10 @@
   color: var(--med-grey);
 }
 
-/* Actions */
-
-.actions {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  justify-content: center;
-  align-items: center;
-  border-top: 1px solid gainsboro;
-  padding-top: 0.875rem;
-  cursor: pointer;
-}
-
-.actions p {
-  padding: 8px 25px;
-  font-weight: 600;
-  color: var(--dark-blue);
-  border-radius: 50px;
-  border: 2px solid var(--dark-blue);
-  width: 100%;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.actions p:hover {
-  background-color: var(--dark-blue);
-  color: var(--white);
-  border-radius: 50px;
-  padding: 8px 25px;
-}
-
-.actions p span {
-  padding-right: 5px;
-}
-
-/* Media Queries */
-@media (min-width: 450px) {
-  .actions {
-    justify-content: space-evenly;
-  }
-  .actions p {
-    width: 45%;
-  }
-}
 
 @media (min-width: 600px) {
-  .info {
-    position: relative;
-  }
+
+
   .book-summary {
     padding: 1.5rem;
   }
@@ -138,6 +75,7 @@
   .summary-title {
     font-size: 1.7rem;
   }
+
   .summary-snippet {
     font-size: 1.1rem;
   }
